+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -1,5 +1,7 @@
 package lotto.domain
 
+import lotto.service.LottoAmountCalculator
+
 class LottoResult(
     private val lottos: List<Lotto>,
     private val winningLotto: WinningLotto,
@@ -16,7 +18,7 @@ class LottoResult(
 
     fun calculateProfitRate(): Double {
         val totalWinningMoney = getRanks().sumOf { it.winningMoney }
-        val totalCost = lottos.size * LottoCalculator.LOTTO_PRICE
+        val totalCost = lottos.size * LottoAmountCalculator.LOTTO_PRICE
         return totalWinningMoney.toDouble() / totalCost
     }
 }
